+++ bft/msg.go
@@ -67,9 +67,9 @@ func (b *BFT) CheckProposerMessage(x *Message) (isPartialQC bool, err lib.ErrorI
 			return false, lib.ErrWrongHeight()
 		}
 		// validate committee height
-		if x.Header.CanopyHeight != b.CanopyHeight {
-			return false, lib.ErrWrongCanopyHeight()
-		}
+		//if x.Header.CanopyHeight != b.CanopyHeight { TODO this shouldn't be here cause it's stopping 'locks' with older Canopy heights from proceeding
+		//	return false, lib.ErrWrongCanopyHeight()
+		//}
 		// sanity check the VRF
 		if err = checkSignatureBasic(x.Vrf); err != nil {
 			return false, err
@@ -105,9 +105,9 @@ func (b *BFT) CheckProposerMessage(x *Message) (isPartialQC bool, err lib.ErrorI
 		if x.Header.Height != b.Height {
 			return false, lib.ErrWrongHeight()
 		}
-		if x.Header.CanopyHeight != b.CanopyHeight {
-			return false, lib.ErrWrongCanopyHeight()
-		}
+		//if x.Header.CanopyHeight != b.CanopyHeight { TODO this shouldn't be here cause it's stopping 'locks' with older Canopy heights from proceeding
+		//	return false, lib.ErrWrongCanopyHeight()
+		//}
 		committeeHeightInState, e := b.LoadCommitteeHeightInState(b.CanopyHeight)
 		if e != nil {
 			return false, e