+++ .docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## [Unreleased] - 2/04/2024
+- Refactor wallet forms and consolidate all of them as one 
+
+## [Unreleased] - 2/03/2024
+- Display UI amounts in CNPY in Wallet/Explorer
+
 ## [Unreleased] - 1/29/2024
 - Add more validator status to wallet
 
@@ -29,7 +35,6 @@
 
 ## [Unreleased] - 1/21/2024
 - Set up prettier for both Wallet and Explorer
->>>>>>> origin/main
 
 ## [Unreleased] - 1/20/2024
 - Wallet UI fixes